moved verifyReddit to tools.py
